`rubocop -A`
+ some manual fixes

+++ db/migrate/20160505142327_remove_package_id_and_profile_id_from_users_if_present.rb
@@ -1,21 +1,15 @@
+# frozen_string_literal: true
+
 class RemovePackageIdAndProfileIdFromUsersIfPresent < ActiveRecord::Migration[4.2]
   def up
-    if column_exists? :users, :package_id
-      remove_column :users, :package_id
-    end
+    remove_column :users, :package_id if column_exists? :users, :package_id
 
-    if column_exists? :users, :profile_id
-      remove_column :users, :profile_id
-    end
+    remove_column :users, :profile_id if column_exists? :users, :profile_id
   end
 
   def down
-    unless column_exists? :users, :package_id
-      add_column :users, :package_id, :integer
-    end
+    add_column :users, :package_id, :integer unless column_exists? :users, :package_id
 
-    unless column_exists? :users, :profile_id
-      add_column :users, :profile_id, :integer
-    end
+    add_column :users, :profile_id, :integer unless column_exists? :users, :profile_id
   end
 end
